docs: Share JSDoc patch with API docs
