+++ config/jsdoc.json
@@ -11,6 +11,7 @@
         "excludePattern": "(^|\\/|\\\\)_"
     },
     "plugins": [
+        "./cli/lib/tsd-jsdoc/patch",
         "./node_modules/jsdoc/plugins/markdown"
     ],
     "templates": {
