+++ src/util/pool.js
@@ -16,7 +16,7 @@ module.exports = pool;
  * @param {number} start Start offset
  * @param {number} end End offset
  * @returns {Uint8Array} Buffer slice
- * @this {Uint8Array}
+ * @this Uint8Array
  */
 
 /**
